us-backend/requirements.txt
uvicorn
asyncpg
FastAPI
validators
PyYAML
aiofiles
python-multipart
Jinja2==3.1.2
